+++ pytsite/plugman/_console_command.py
@@ -1,6 +1,6 @@
-"""PytSite Plugin Manager Console Commands 
+"""PytSite Plugin Manager Console Commands
 """
-from pytsite import console as _console, reg as _reg, reload as _reload
+from pytsite import console as _console, reg as _reg, reload as _reload, lang as _lang
 from . import _api, _error
 
 __author__ = 'Alexander Shepetko'
@@ -40,6 +40,8 @@ class Install(_console.Command):
                 if not _api.is_installed(plugin_name):
                     _api.install(plugin_name)
 
+        _reload.reload()
+
 
 class Upgrade(_console.Command):
     """plugman:upgrade
@@ -48,6 +50,7 @@ class Upgrade(_console.Command):
     def __init__(self):
         super().__init__()
 
+        self._define_option(_console.option.Bool('reload', default=True))
         self._define_argument(_console.argument.Argument('plugin_name'))
 
     @property
@@ -59,17 +62,12 @@ class Upgrade(_console.Command):
         return 'pytsite.plugman@console_command_description_upgrade'
 
     def execute(self):
-        plugin_name = self.get_argument_value(0)
+        try:
+            _console.print_info(_lang.t('pytsite.plugman@upgrading_plugins'))
+            _api.upgrade(self.get_argument_value(0))
 
-        if plugin_name:
-            try:
-                _api.upgrade(plugin_name)
-            except _error.PluginNotInstalled as e:
-                raise _console.error.Error(e)
-        else:
-            # Upgrade all installed plugins
-            for name, info in _api.get_plugin_info().items():
-                if info.get('installed_version'):
-                    _api.upgrade(name)
+        except _error.PluginNotInstalled as e:
+            raise _console.error.Error(e)
 
-        _reload.reload()
+        if self.get_option_value('reload'):
+            _reload.reload()