+++ pytsite/theme/__init__.py
@@ -18,7 +18,7 @@ def _init():
     lang.register_package(__name__)
 
     assetman.register_package(__name__)
-    assetman.t_js(__name__ + '@**/*.js')
+    assetman.t_js(__name__ + '@**')
 
     # App's logo URL resolver
     def logo_url(width: int = 50, height: int = 50):